# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
  description: Sakani Dz - Real Estate & Tourism Platform API
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        preferredWilaya:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    VendorSummary:
      type: object
      required: [id, firstName, lastName, phone]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        isVerified:
          type: boolean
...
    Property:
      type: object
      required: [id, vendorId, type, price, wilaya, status, photos, equipment, createdAt]
      properties:
        id:
          type: integer
        vendorId:
          type: integer
        vendor:
          $ref: "#/components/schemas/VendorSummary"
        type:
          type: string
          enum: [apartment, villa]
        apartmentType:
          type: ["string", "null"]
          enum: [F1, F2, F3, F4, F5, null]
        description:
          type: ["string", "null"]
        equipment:
          type: array
          items:
            type: string
        price:
          type: number
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, rented]
        wilaya:
          type: string
        siteId:
          type: ["integer", "null"]
        views:
          type: integer
        isFeatured:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    TouristSpot:
      type: object
      required: [id, name, wilaya, photos, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        wilaya:
          type: string
        description:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        photos:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    Favorite:
      type: object
      required: [id, clientId, propertyId, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        propertyId:
          type: integer
        property:
          $ref: "#/components/schemas/Property"
        createdAt:
          type: string
          format: date-time
    HistoryEntry:
      type: object
      required: [id, clientId, entryType, viewedAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        propertyId:
          type: ["integer", "null"]
        touristSpotId:
          type: ["integer", "null"]
        entryType:
          type: string
          enum: [property, tourist_spot]
        property:
          $ref: "#/components/schemas/Property"
        touristSpot:
          $ref: "#/components/schemas/TouristSpot"
        viewedAt:
          type: string
          format: date-time
...
    FavoriteListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Favorite"
    HistoryListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
...
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        phone:
          type: string
        preferredWilaya:
          type: string
...
    FavoriteInput:
      type: object
      required: [propertyId]
      properties:
        propertyId:
          type: integer
    HistoryInput:
      type: object
      required: [entryType]
      properties:
        propertyId:
          type: integer
        touristSpotId:
          type: integer
        entryType:
          type: string
          enum: [property, tourist_spot]
...
paths:
...
  /clients/me:
    get:
      operationId: getClientProfile
      tags: [clients]
      summary: Get current client profile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClientProfile
      tags: [clients]
      summary: Update client profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # â”€â”€ Favorites â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
  /clients/me/favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List client favorites
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteListResponse"
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add a property to favorites
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorite"
  /clients/me/favorites/{propertyId}:
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a property from favorites
      security:
        - BearerAuth: []
      parameters:
        - name: propertyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # â”€â”€ History â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
  /clients/me/history:
    get:
      operationId: listHistory
      tags: [history]
      summary: List client browsing history
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryListResponse"
    post:
      operationId: addHistoryEntry
      tags: [history]
      summary: Record a history entry
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HistoryInput"
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryEntry"
  /clients/me/history/{id}:
    delete:
      operationId: deleteHistoryEntry
      tags: [history]
      summary: Delete a history entry
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # â”€â”€ Properties â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...