# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Sakani Dz - Real Estate & Tourism Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { PropertyApartmentType } from './propertyApartmentType';
import type { PropertyStatus } from './propertyStatus';
import type { PropertyType } from './propertyType';
import type { VendorSummary } from './vendorSummary';

export interface Property {
  id: number;
  vendorId: number;
  vendor?: VendorSummary;
  type: PropertyType;
  /** @nullable */
  apartmentType?: PropertyApartmentType;
  /** @nullable */
  description?: string | null;
  equipment: string[];
  price: number;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  photos: string[];
  status: PropertyStatus;
  wilaya: string;
  /** @nullable */
  siteId?: number | null;
  views?: number;
  isFeatured?: boolean;
  createdAt: Date;
}